requirements.txt
botogram2
django
psycopg2-binary
pillow  # required by django ImageField
djangorestframework
